Main view integrated

+++ WeatherForecast/DomainLayer/WeatherForecastOperations/Models/InnerDetailModels/City.swift
@@ -9,13 +9,13 @@ import Foundation
 
 // MARK: - City
 class City: CodableDataProtocol {
-    let id: Int?
+    let id: Int64?
     let name: String?
     let coord: Coordinates?
     let country: String?
     let population, timezone: Int?
 
-    init(id: Int?, name: String?, coord: Coordinates?, country: String?, population: Int?, timezone: Int?) {
+    init(id: Int64?, name: String?, coord: Coordinates?, country: String?, population: Int?, timezone: Int?) {
         self.id = id
         self.name = name
         self.coord = coord
